protect patient routes

+++ resources/patients/routes.py
@@ -1,10 +1,11 @@
 from flask.views import MethodView
 from flask_smorest import abort
 from sqlalchemy.exc import IntegrityError
-from schemas import InterventionSchema, UpdatePatientSchema, PatientSchema, DeletePatientSchema, PatientSchemaNested
+from schemas import InterventionSchema, UpdatePatientSchema, PatientSchema, AuthPatientSchema, PatientSchemaNested
 from . import bp
 from .PatientModel import PatientModel
 from ..therapists.TherapistModel import TherapistModel
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 from db import patients, interventions
 
@@ -25,16 +26,17 @@ class PatientList(MethodView):
     #     patients[uuid4().hex] = patient_data
     #     return patient_data, 201
     
-    @bp.arguments(PatientSchema)
-    @bp.response(201, PatientSchema)
-    def post(self, patient_data):
-        patient = PatientModel()
-        try:
-            patient.from_dict(patient_data)
-            patient.save()
-            return patient_data
-        except IntegrityError:
-            abort(400, message='Patient is already in medical record system.')
+    # WE MOVED THIS TO TO AUTH_ROUTES
+    # @bp.arguments(PatientSchema)
+    # @bp.response(201, PatientSchema)
+    # def post(self, patient_data):
+    #     patient = PatientModel()
+    #     try:
+    #         patient.from_dict(patient_data)
+    #         patient.save()
+    #         return patient_data
+    #     except IntegrityError:
+    #         abort(400, message='Patient is already in medical record system.')
 
 # @app.delete('/patient/<patient_id>')
 # def discharge_patient(patient_id):
@@ -45,14 +47,30 @@ class PatientList(MethodView):
 #         return {'message': 'patient not found'}, 400
 #     return {'message':f"{store} has been discharged and deleted from database"}, 202
 # ************************************************************************************
-    @bp.arguments(DeletePatientSchema)
+    @jwt_required()
+    @bp.arguments(AuthPatientSchema)
     def delete(self, patient_data):
-        patient = PatientModel.query.filter_by(id=patient_data['id']).first()
+        patient_id = get_jwt_identity()
+        patient = PatientModel.query.get(patient_id)
         # .first() gives us the first value. you can also give .all() which would return a list, but because there is only 1 unique username or user_id, you would have to then index into the list which is just extra step.
-        if patient and patient.last_name == patient_data['last_name']:
+        if patient and patient.first_name == patient_data['first_name'] and patient.last_name == patient_data['last_name'] and patient.id == patient_data['id']:
             patient.delete()
             return {'message':f'{patient_data["first_name"]} {patient_data["last_name"]} has been discharged and removed from medical record system'}, 202
         abort(400, message='Patient ID Invalid')
+
+    @jwt_required()
+    @bp.arguments(UpdatePatientSchema)
+    @bp.response(202, PatientSchema)
+    def put(self, patient_data):
+        patient_id = get_jwt_identity()
+        patient = PatientModel.query.get_or_404(patient_id, description="Patient Not Found")
+        if patient:
+            try:
+                patient.from_dict(patient_data)
+                patient.save()
+                return patient
+            except IntegrityError:
+                abort(400, message='Patient already exists in system.')
 
 @bp.route('/patient/<patient_id>')
 class Patient(MethodView):
@@ -69,17 +87,6 @@ class Patient(MethodView):
 #     except KeyError:
 #         return {'message':'patient not found'}, 400
 
-    @bp.arguments(UpdatePatientSchema)
-    @bp.response(202, PatientSchema)
-    def put(self, patient_data, patient_id):
-        patient = PatientModel.query.get_or_404(patient_id, description="Patient Not Found")
-        if patient:
-            try:
-                patient.from_dict(patient_data)
-                patient.save()
-                return patient
-            except IntegrityError:
-                abort(400, message='Patient already exists in system.')
 # @app.put('/patient/<patient_id>')
 # def update_patient(patient_id):
 #     patient_data = request.get_json()
